Fix medium-severity static analyzer issues
- Add explicit != 0 comparisons to strcmp/strncmp/strcasecmp calls
- Add explicit casts for multi-level pointer conversions
- Fix integer division precision loss in fight.c patience calculation

+++ code/fight.c
@@ -1682,7 +1682,7 @@ static void develop_hatred(struct char_data* ch, struct char_data* v) {
   diff /= 20;
 
   if (GET_MAX_HIT(ch)) {
-    patience = 100 * (float)(GET_HIT(ch) / GET_MAX_HIT(ch));
+    patience = (int)(100.0f * (float)GET_HIT(ch) / (float)GET_MAX_HIT(ch));
   } else {
     patience = 10;
   }
